requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0

# HTTP Requests
httpx>=0.28.0
requests>=2.32.0
aiohttp>=3.11.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# ML/AI
scikit-learn>=1.4.0
joblib>=1.3.2

# Threat Intelligence APIs
vt-py==0.18.4  # VirusTotal
OTXv2==1.5.12  # AlienVault OTX

# Enrichment Services
python-whois==0.8.0
dnspython<=2.0.0
geoip2==4.7.0
ipwhois==1.2.0

# Security & Validation
email-validator==2.1.0
python-multipart==0.0.6

# Scheduling
apscheduler==3.10.4

# Notifications
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
